+++ ModuleLowpassFilter.h
@@ -24,7 +24,7 @@ class ModuleLowpassFilter : public Module
     
     // Methods
     ModuleLowpassFilter();
-    uint32_t compute();
+    uint16_t compute();
     
     // Inputs
     Module *audio_input;
