+++ ModuleMap.cpp
@@ -9,7 +9,7 @@ ModuleMap::ModuleMap(uint32_t low, uint32_t high)
   this->high = high;
 }
 
-uint32_t ModuleMap::compute()
+uint16_t ModuleMap::compute()
 {  
   return(map(this->readInput(input), 0, 4095, low, high));
 }
